refactor: Refactor package structure for loan application router
