+++ infrastructure/entry-points/reactive-web/src/test/java/co/com/crediya/api/v1/LoanApplicationRouterV1Test.java
@@ -5,8 +5,8 @@ import co.com.crediya.api.dto.loanapplication.ShortLoanApplicationDTO;
 import co.com.crediya.api.mapper.LoanApplicationDTOMapper;
 import co.com.crediya.api.presentation.contract.DTOValidator;
 import co.com.crediya.api.presentation.contract.UUIDValidator;
-import co.com.crediya.api.presentation.contract.loanapplication.v1.LoanApplicationRouterV1;
-import co.com.crediya.api.presentation.contract.loanapplication.v1.handler.RegisterLoanApplicationHandlerV1;
+import co.com.crediya.api.presentation.loanapplication.v1.LoanApplicationRouterV1;
+import co.com.crediya.api.presentation.loanapplication.v1.handler.RegisterLoanApplicationHandlerV1;
 import co.com.crediya.model.loanapplication.LoanApplication;
 import co.com.crediya.model.loantype.enums.LoanTypes;
 import co.com.crediya.model.status.enums.Statuses;
